--- anabada/src/main/java/co/anabada/common/TestControl.java
@@ -1,19 +1,0 @@
-package co.anabada.common;
-
-import java.io.IOException;
-
-import javax.servlet.RequestDispatcher;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-public class TestControl implements Control {
-
-	@Override
-	public void exec(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		String path = "main/test.tiles";
-		RequestDispatcher dispatch = req.getRequestDispatcher(path);
-		dispatch.forward(req, resp);
-	}
-
-}
