+++ anabada/src/main/webapp/WEB-INF/layout/template.jsp
@@ -12,6 +12,9 @@
 <meta content="" name="keywords">
 <meta content="" name="description">
 
+<!-- jQuery -->
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.7.1/jquery.min.js"></script>
+
 <!-- Google Web Fonts -->
 <link rel="preconnect" href="https://fonts.googleapis.com">
 <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
