+++ anabada/src/test/java/co/anabada/MapperTest.java
@@ -10,13 +10,10 @@ public class MapperTest {
 	public static void main(String[] args) {
 		// mapper 테스트.
 		MemberMapper mapper = DataSource.getInstance().openSession(true).getMapper(MemberMapper.class);
-		
+
 		List<Member> list = mapper.memberList();
-		for(Member m : list) {
+		for (Member m : list) {
 			System.out.println(m.toString());
 		}
-		
-		
-
 	}
 }
