+++ anabada/src/test/resources/config/db.properties
@@ -1,4 +1,4 @@
 driver=oracle.jdbc.OracleDriver
 url=jdbc:oracle:thin:@localhost:1521:xe
-username=yedam
-password=yedam
+username=bada
+password=bada
